% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_conv_1d.R
\name{layer_residual_conv_1d}
\alias{layer_residual_conv_1d}
\title{Wavenet residual connections}
\usage{
layer_residual_conv_1d(object, filters, kernel_size, dilation_rate,
  name = NULL, trainable = TRUE)
}
\arguments{
\item{object}{Model or layer object}

\item{filters}{Integer, the dimensionality of the output space (i.e. the
number of output filters in the convolution).}

\item{kernel_size}{An integer or list of a single integer, specifying the
length of the 1D convolution window.}

\item{dilation_rate}{an integer or list of a single integer, specifying the
dilation rate to use for dilated convolution. Currently, specifying any
\code{dilation_rate} value != 1 is incompatible with specifying any \code{strides}
value != 1.}

\item{name}{An optional name string for the layer. Should be unique in a
model (do not reuse the same name twice). It will be autogenerated if it
isn't provided.}

\item{trainable}{Whether the layer weights will be updated during training.}
}
\description{
Residual connection as described in section 2.3 of
\href{https://arxiv.org/abs/1609.03499}{van der Oord et al., \cite{WaveNet: A Generative Model for Raw Audio}}.
}
