% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mu_law}
\alias{mu_law}
\title{Mu Law encodes the audio}
\usage{
mu_law(x, mu = 255)
}
\arguments{
\item{x}{Tensor representing an audio.}

\item{mu}{quantization constant.}
}
\description{
See section 2.2 of \href{https://arxiv.org/abs/1609.03499}{van der Oord et al., \cite{WaveNet: A Generative Model for Raw Audio}}.
for more information.
}
