% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mu_law_and_one_hot_encode}
\alias{mu_law_and_one_hot_encode}
\title{Makes the transformation proposed in WaveNet and One Hot encodes the output}
\usage{
mu_law_and_one_hot_encode(x, mu)
}
\arguments{
\item{x}{Tensor representing an audio.}

\item{mu}{quantization constant.}
}
\description{
See section 2.2 of \href{https://arxiv.org/abs/1609.03499}{van der Oord et al., \cite{WaveNet: A Generative Model for Raw Audio}}.
for more information.
}
