% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_crop}
\alias{random_crop}
\title{Takes a random crop of an audio Tensor}
\usage{
random_crop(x, seconds)
}
\arguments{
\item{x}{An audio file decode by `tensorflow::tf$audio$decode_wav`.}

\item{seconds}{The number of seconds of the random crop.}
}
\description{
Takes a random crop of an audio Tensor
}
