% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavenet.R
\name{wavenet}
\alias{wavenet}
\title{Builds the wavenet model}
\usage{
wavenet(filters = 16, kernel_size = 2, residual_blocks,
  input_shape = list(NULL, 1), input_tensor = NULL,
  initial_kernel_size = 32, initial_filters = 32,
  output_channels = 256, output_activation = "softmax")
}
\arguments{
\item{filters}{Integer, the dimensionality of the output space (i.e. the
number of output filters in the convolution).}

\item{kernel_size}{An integer or list of a single integer, specifying the
length of the 1D convolution window.}

\item{residual_blocks}{if a single integer: the number of residual blocks in the
model - the dilation rate of each block is calculated by `2^i`. if a vector,
then it's used as the dilation rate of each residual block.}

\item{input_shape}{input shape for the model (not including the axis dimension).
Typically lenght 2 numeric vector. Used to build the `input_tensor` if no
`input_tensor` is specified.}

\item{input_tensor}{Tensor to use as input for the model. Usually a 3d tensor.}

\item{initial_kernel_size}{kernel size of the first causal convolution.}

\item{initial_filters}{number of filters in the initial causal convolution.}

\item{output_channels}{number of channels in the output.}

\item{output_activation}{activation function for the last layer. default: 'softmax'}
}
\description{
Builds the wavenet model as described in
\href{https://arxiv.org/abs/1609.03499}{van der Oord et al., \cite{WaveNet: A Generative Model for Raw Audio}}.
}
